feat: enhance data explorer URL functionality with new builders and configuration updates
