+++ statgpt/common/data/sdmx/common/data_explorer_url.py
@@ -1,19 +1,53 @@
-"""Configuration for building public web *data explorer* deep links (SDMX-backed queries).
+"""Configuration and builders for data explorer deep links (SDMX-backed queries)."""
 
-Default behaviour matches the original product: a dataflow URN as one query parameter, a
-**filter** value produced from the DSD key as an SDMX-style key string, and
-``startPeriod``/``endPeriod`` when the portal expects SDMX 2.1 time parameters.
-"""
-
-from typing import Literal
+import logging
+from collections.abc import Callable
+from enum import StrEnum
+from typing import Any
+from urllib.parse import quote, urlencode
 
 from pydantic import Field, model_validator
 
 from statgpt.common.data.base import BaseModel
 
-TimeEncodingSdmx = Literal['start_end_sdmx21', 'none', 'in_aggregated_filter']
-FilterFormatSdmx = Literal['sdmx_key_string', 'key_value_aggregated']
-AggregatedValueModeSdmx = Literal['code', 'name']
+
+class TimeEncodingSdmx(StrEnum):
+    """Where the time range is encoded in the data explorer URL."""
+
+    start_end_sdmx21 = "start_end_sdmx21"
+    """Use separate query params (`startPeriod`, `endPeriod`)."""
+
+    none = "none"
+    """Do not add explicit time to the URL."""
+
+    in_aggregated_filter = "in_aggregated_filter"
+    """Put time into the aggregated filter entry."""
+
+
+class FilterFormatSdmx(StrEnum):
+    """How categorical filters are serialized in the URL."""
+
+    sdmx_key_string = "sdmx_key_string"
+    """Use positional SDMX key format."""
+
+    key_value_aggregated = "key_value_aggregated"
+    """Use `NAME=value` entries joined by delimiters."""
+
+
+class AggregatedValueModeSdmx(StrEnum):
+    """How each aggregated filter value is represented."""
+
+    code = "code"
+    """Keep raw SDMX code values."""
+
+    name = "name"
+    """Resolve SDMX codes to display labels where available."""
+
+
+DimensionValuesResolver = Callable[[str, list[str]], list[str]]
+SdmxKeyBuilder = Callable[[Any, dict[str, list[str]]], str]
+
+_log = logging.getLogger(__name__)
 
 
 class DataExplorerUrlConfig(BaseModel):
@@ -48,7 +82,7 @@ class DataExplorerUrlConfig(BaseModel):
         description="Query parameter name for the dataflow URN when includeDataflowUrnParam is true.",
     )
     filter_format: FilterFormatSdmx = Field(
-        default="sdmx_key_string",
+        default=FilterFormatSdmx.sdmx_key_string,
         description="How the filter value is built from the query key and DSD.",
     )
     include_series_key_filter: bool = Field(
@@ -60,7 +94,7 @@ class DataExplorerUrlConfig(BaseModel):
         description="Query parameter name for the series filter string.",
     )
     time_encoding: TimeEncodingSdmx = Field(
-        default="start_end_sdmx21",
+        default=TimeEncodingSdmx.start_end_sdmx21,
         description=(
             "start_end_sdmx21: add startPeriod/endPeriod. "
             "none: no separate time query parameters. "
@@ -107,13 +141,152 @@ class DataExplorerUrlConfig(BaseModel):
 
     @model_validator(mode='after')
     def _aggregated_time_consistent(self) -> "DataExplorerUrlConfig":
-        if self.time_encoding == "in_aggregated_filter" and not self.aggregated_time_param:
+        if (
+            self.time_encoding == TimeEncodingSdmx.in_aggregated_filter
+            and not self.aggregated_time_param
+        ):
             raise ValueError(
                 "aggregatedTimeParam is required when timeEncoding is in_aggregated_filter"
             )
         if (
-            self.time_encoding == "in_aggregated_filter"
-            and self.filter_format != "key_value_aggregated"
+            self.time_encoding == TimeEncodingSdmx.in_aggregated_filter
+            and self.filter_format != FilterFormatSdmx.key_value_aggregated
         ):
             raise ValueError("in_aggregated_filter requires filterFormat key_value_aggregated")
         return self
+
+
+def _dataflow_component_ids_in_order(dsd: Any) -> list[str]:
+    try:
+        return [c.id for c in dsd.dimensions.components]
+    except (AttributeError, TypeError):
+        return []
+
+
+def _ordered_dimension_keys(dsd: Any, key_dict: dict[str, list[str]]) -> list[str]:
+    order = _dataflow_component_ids_in_order(dsd)
+    pos: dict[str, int] = {d: i for i, d in enumerate(order)}
+
+    def _sort_key(dim_id: str) -> tuple[int, str]:
+        return (pos.get(dim_id, 10_000), dim_id)
+
+    return sorted(key_dict.keys(), key=_sort_key)
+
+
+def _build_aggregated_filter_value(
+    dsd: Any,
+    key_dict: dict[str, list[str]],
+    sdmx_query: Any,
+    cfg: DataExplorerUrlConfig,
+    dimension_values_resolver: DimensionValuesResolver | None = None,
+) -> str:
+    dsep, kvsep, vsep = (
+        cfg.aggregated_entry_delimiter,
+        cfg.aggregated_key_value_separator,
+        cfg.aggregated_values_separator,
+    )
+    name_map = cfg.aggregated_dimension_param_names
+    parts: list[str] = []
+    for dim_id in _ordered_dimension_keys(dsd, key_dict):
+        codes = key_dict[dim_id]
+        if not codes:
+            continue
+        label = name_map.get(dim_id, dim_id)
+        values = list(codes)
+        if (
+            cfg.aggregated_dimension_value_mode.get(dim_id) == AggregatedValueModeSdmx.name
+            and dimension_values_resolver is not None
+        ):
+            values = dimension_values_resolver(dim_id, codes)
+        value = vsep.join(values)
+        parts.append(f"{label}{kvsep}{value}")
+
+    if (
+        cfg.time_encoding == TimeEncodingSdmx.in_aggregated_filter
+        and cfg.aggregated_time_param
+        and getattr(sdmx_query, "time_dimension_query", None)
+    ):
+        tq = sdmx_query.time_dimension_query
+        sp, ep = tq.start_period, tq.end_period
+        tsep = cfg.aggregated_time_range_separator
+        if sp and ep:
+            parts.append(f"{cfg.aggregated_time_param}{kvsep}{sp}{tsep}{ep}")
+        elif sp or ep:
+            one = (sp or ep) or ""
+            parts.append(f"{cfg.aggregated_time_param}{kvsep}{one}")
+    if not parts:
+        return ""
+    return dsep.join(parts)
+
+
+def build_data_explorer_url_query(
+    base_url: str,
+    short_urn: str,
+    dsd: Any,
+    key_dict: dict[str, list[str]],
+    sdmx_query: Any,
+    config: DataExplorerUrlConfig | None,
+    sdmx_key_builder: SdmxKeyBuilder | None = None,
+    dimension_values_resolver: DimensionValuesResolver | None = None,
+) -> str:
+    """Return a full explorer URL: ``base`` + ``?`` + query (legacy if ``config`` is None)."""
+    cfg = config or DataExplorerUrlConfig()
+    params: dict[str, str] = {}
+
+    if cfg.include_dataflow_urn_param:
+        params[cfg.dataset_urn_param] = short_urn
+
+    if cfg.include_series_key_filter:
+        if cfg.filter_format == FilterFormatSdmx.sdmx_key_string:
+            if sdmx_key_builder is None:
+                raise ValueError("sdmx_key_builder is required for filterFormat='sdmx_key_string'")
+            params[cfg.series_key_filter_param] = sdmx_key_builder(dsd, key_dict)
+        else:
+            filter_body = _build_aggregated_filter_value(
+                dsd=dsd,
+                key_dict=key_dict,
+                sdmx_query=sdmx_query,
+                cfg=cfg,
+                dimension_values_resolver=dimension_values_resolver,
+            )
+            if filter_body:
+                params[cfg.series_key_filter_param] = filter_body
+
+    if cfg.time_encoding == TimeEncodingSdmx.start_end_sdmx21:
+        _add_sdmx21_time_params(params, sdmx_query)
+    # in_aggregated_filter: time is inside the filter; none: no time
+
+    safe_chars = "+,.:*" if cfg.filter_format == FilterFormatSdmx.sdmx_key_string else ",.:*"
+    query_string = urlencode(params, quote_via=quote, safe=safe_chars)
+    if not query_string:
+        return base_url
+    return f"{base_url}?{query_string}"
+
+
+def _add_sdmx21_time_params(params: dict[str, str], sdmx_query: Any) -> None:
+    try:
+        if td_query := getattr(sdmx_query, "time_dimension_query", None):
+            start_period = td_query.start_period
+            end_period = td_query.end_period
+        else:
+            start_period = None
+            end_period = None
+
+        if start_period:
+            params['startPeriod'] = start_period if '-' in start_period else f"{start_period}-A"
+        if end_period:
+            params['endPeriod'] = end_period if '-' in end_period else f"{end_period}-A"
+    except Exception as e:  # noqa: BLE001
+        _log.exception(e)
+
+
+def build_data_explorer_dataset_url(
+    base_url: str, short_urn: str, config: DataExplorerUrlConfig | None
+) -> str:
+    """Build a link for ``dataset_url`` when ``useDataExplorerForDatasetUrl`` is set."""
+    cfg = config or DataExplorerUrlConfig()
+    if not cfg.include_dataflow_urn_param:
+        return base_url.rstrip("?&")
+    param = {cfg.dataset_urn_param: short_urn}
+    query_string = urlencode(param, quote_via=quote, safe="+,.:*()")
+    return f"{base_url}?{query_string}"
